+++ app/Http/Controllers/ServiceReportsController.php
@@ -50,33 +50,31 @@ class ServiceReportsController extends Controller
      */
     public function create()
     {
-        $sales = User::where('role', 'Sales')->get();
-        $clients = Client::all();
+        $sales = User::where('role', 'Sales')->select('id', 'name')->get();
+        $clients = Client::select('id', 'company', 'id_sales')->with('sales:id,name')->get();
         $dateNow = Carbon::now();
         $numberS = Reports::whereYear('date', $dateNow)->where('id_technician', Auth::user()->id)->count();
         $formattedNumberS = str_pad($numberS + 1, 3, '0', STR_PAD_LEFT);
         $monthNow = $dateNow->month;
         $formattedMonthNow = $this->convertToRoman($monthNow);
-        $pic = Pic::join('client as c', 'c.id', '=', 'pic.id_client')->select('pic.*')->get();
-        // default selections
+        
         $selectedSalesId = 23;
         $selectedClientId = 5568;
         $selectedPICId = null;
         $selectedMachineId = null;
 
-        return view('pages.technician.service-reports.form', compact('sales', 'pic', 'formattedNumberS', 'formattedMonthNow', 'clients', 'selectedSalesId', 'selectedClientId', 'selectedPICId', 'selectedMachineId'));
+        return view('pages.technician.service-reports.form', compact('sales', 'clients', 'formattedNumberS', 'formattedMonthNow', 'selectedSalesId', 'selectedClientId', 'selectedPICId', 'selectedMachineId'));
     }
 
     public function createByUnit($id_unit)
     {
-        $sales = User::where('role', 'Sales')->get();
-        $clients = Client::all();
+        $sales = User::where('role', 'Sales')->select('id', 'name')->get();
+        $clients = Client::select('id', 'company', 'id_sales')->with('sales:id,name')->get();
         $dateNow = Carbon::now();
         $numberS = Reports::whereYear('date', $dateNow)->where('id_technician', Auth::user()->id)->count();
         $formattedNumberS = str_pad($numberS + 1, 3, '0', STR_PAD_LEFT);
         $monthNow = $dateNow->month;
         $formattedMonthNow = $this->convertToRoman($monthNow);
-        $pic = Pic::join('client as c', 'c.id', '=', 'pic.id_client')->select('pic.*')->get();
 
         $machine = \App\Models\Machine::where('id_unit', $id_unit)->first();
         if (!$machine) {
@@ -90,10 +88,9 @@ class ServiceReportsController extends Controller
 
         return view('pages.technician.service-reports.form', compact(
             'sales',
-            'pic',
+            'clients',
             'formattedNumberS',
             'formattedMonthNow',
-            'clients',
             'selectedSalesId',
             'selectedClientId',
             'selectedPICId',
@@ -103,16 +100,15 @@ class ServiceReportsController extends Controller
 
     public function createByUnitMachine($id_unit, $id_machine)
     {
-        $sales = User::where('role', 'Sales')->get();
-        $clients = Client::all();
+        $sales = User::where('role', 'Sales')->select('id', 'name')->get();
+        $clients = Client::select('id', 'company', 'id_sales')->with('sales:id,name')->get();
         $dateNow = Carbon::now();
         $numberS = Reports::whereYear('date', $dateNow)->where('id_technician', Auth::user()->id)->count();
         $formattedNumberS = str_pad($numberS + 1, 3, '0', STR_PAD_LEFT);
         $monthNow = $dateNow->month;
         $formattedMonthNow = $this->convertToRoman($monthNow);
-        $pic = Pic::join('client as c', 'c.id', '=', 'pic.id_client')->select('pic.*')->get();
 
-        $machine = \App\Models\Machine::find($id_machine);
+        $machine = \App\Models\Machine::with('unit')->find($id_machine);
         if (!$machine) {
             abort(404);
         }
@@ -122,6 +118,8 @@ class ServiceReportsController extends Controller
             $id_unit = $machine->id_unit;
         }
 
+        $isInternalStock = ($machine->id_client == 5387 || $machine->id_client == 1277);
+
         $selectedClientId = $machine->id_client ?? 5568;
         $selectedSalesId = optional(Client::find($selectedClientId))->id_sales ?? 23;
         $selectedPICId = optional(Pic::where('id_client', $selectedClientId)->first())->id;
@@ -129,27 +127,36 @@ class ServiceReportsController extends Controller
 
         return view('pages.technician.service-reports.form', compact(
             'sales',
-            'pic',
+            'clients',
             'formattedNumberS',
             'formattedMonthNow',
-            'clients',
             'selectedSalesId',
             'selectedClientId',
             'selectedPICId',
-            'selectedMachineId'
+            'selectedMachineId',
+            'isInternalStock',
+            'machine'
         ));
+    }
+
+    public function indexByMachine($id_machine)
+    {
+        $machine = \App\Models\Machine::with(['unit', 'client'])->find($id_machine);
+        if (!$machine) {
+            abort(404);
+        }
+        return view('pages.technician.service-reports.machine-index', compact('machine'));
     }
 
     public function createByMachine($id_machine)
     {
-        $sales = User::where('role', 'Sales')->get();
-        $clients = Client::all();
+        $sales = User::where('role', 'Sales')->select('id', 'name')->get();
+        $clients = Client::select('id', 'company', 'id_sales')->with('sales:id,name')->get();
         $dateNow = Carbon::now();
         $numberS = Reports::whereYear('date', $dateNow)->where('id_technician', Auth::user()->id)->count();
         $formattedNumberS = str_pad($numberS + 1, 3, '0', STR_PAD_LEFT);
         $monthNow = $dateNow->month;
         $formattedMonthNow = $this->convertToRoman($monthNow);
-        $pic = Pic::join('client as c', 'c.id', '=', 'pic.id_client')->select('pic.*')->get();
 
         $machine = \App\Models\Machine::find($id_machine);
         if (!$machine) {
@@ -163,15 +170,33 @@ class ServiceReportsController extends Controller
 
         return view('pages.technician.service-reports.form', compact(
             'sales',
-            'pic',
+            'clients',
             'formattedNumberS',
             'formattedMonthNow',
-            'clients',
             'selectedSalesId',
             'selectedClientId',
             'selectedPICId',
             'selectedMachineId'
         ));
+    }
+
+    public function dataByMachine($id_machine)
+    {
+        $data = Reports::with('technician')
+            ->where('id_machine', $id_machine)
+            ->orderByDesc('date')
+            ->get(['id', 'no_service', 'type', 'jobdesc', 'date', 'id_technician']);
+
+        return response()->json(['data' => $data->map(function ($r) {
+            return [
+                'id'         => $r->id,
+                'no_service' => $r->no_service,
+                'type'       => $r->type,
+                'jobdesc'    => $r->jobdesc,
+                'date'       => $r->date,
+                'technician' => optional($r->technician)->name ?? '-',
+            ];
+        })]);
     }
 
     /**
@@ -206,6 +231,7 @@ class ServiceReportsController extends Controller
         $reports->id_machine = $request->machine;
         $reports->no_service = $request->no_service;
         $reports->type = $request->type;
+        $reports->pm_level = $request->pm_level;
         $reports->running = $request->running;
         $reports->load = $request->load;
         $reports->date = $request->date;
@@ -243,7 +269,8 @@ class ServiceReportsController extends Controller
      */
     public function edit($id)
     {
-        $sales = User::where('role', 'Sales')->get();
+        $sales = User::where('role', 'Sales')->select('id', 'name')->get();
+        $clients = Client::select('id', 'company', 'id_sales')->with('sales:id,name')->get();
         $report = Reports::find($id);
         $image = ReportsPict::where('id_reports', $id)->get();
         $dateNow = Carbon::now();
@@ -251,10 +278,8 @@ class ServiceReportsController extends Controller
         $formattedNumberS = str_pad($numberS + 1, 3, '0', STR_PAD_LEFT);
         $monthNow = $dateNow->month;
         $formattedMonthNow = $this->convertToRoman($monthNow);
-        $clients = Client::all();
-        $pic = Pic::all();
-        // dd($image);
-        return view('pages.technician.service-reports.form', compact('sales','pic', 'formattedNumberS', 'formattedMonthNow', 'report', 'image', 'clients'));
+        
+        return view('pages.technician.service-reports.form', compact('sales', 'clients', 'formattedNumberS', 'formattedMonthNow', 'report', 'image'));
     }
 
     /**
@@ -284,8 +309,9 @@ class ServiceReportsController extends Controller
         // Masukan Data ke Service Reports
         $reports = Reports::find($id);
         // $reports->id_technician = $request->technician;
-        // $reports->id_pic = $request->id_pic;
+        $reports->id_pic = $request->id_pic;
         $reports->type = $request->type;
+        $reports->pm_level = $request->pm_level;
         $reports->id_machine = $request->machine;
         // $reports->no_service = $request->no_service;
         $reports->running = $request->running;
@@ -313,8 +339,10 @@ class ServiceReportsController extends Controller
         $pic = ReportsPict::where('id_reports', $id)->get();
 
         $delService = $service->delete();
+
+        $delPict = true;
         foreach ($pic as $pict) {
-            $delPict = $pict->delete();
+            $delPict = $pict->delete() && $delPict;
         }
 
         if ($delService && $delPict) {
@@ -411,12 +439,104 @@ class ServiceReportsController extends Controller
         }
     }
 
+    public function inputImageV2(Request $request, $id)
+    {
+        $images = $request->file('image', []);
+        $descriptions = $request->input('description', []);
+
+        $status = false;
+        foreach ($images as $item => $foto) {
+            $photo = new ReportsPict();
+            $photo->id_reports = $id;
+            $photo->keterangan = $descriptions[$item] ?? '';
+
+            $image_ext = $foto->getClientOriginalExtension();
+            $image_name = Str::random(8);
+
+            $upload_path = 'asset/reports';
+            $imagename = $upload_path . '/' . $image_name . '.' . $image_ext;
+
+            // Crop otomatis jadi square, portrait/landscape apapun mengikuti
+            $img = Image::make($foto->path());
+            $img->fit(600, 600);
+            $img->save($imagename);
+
+            $photo->picture = $imagename;
+            $status = $photo->save();
+        }
+
+        if ($status) {
+            return redirect('service-reports/' . $id)->with('success', 'Data Has been created');
+        }
+    }
+
     public function deleteImage($id)
     {
         $photos = ReportsPict::where('id_reports', $id)->get();
+
+        $status = true;
         foreach ($photos as $picture) {
-            $status = $picture->delete();
+            $status = $picture->delete() && $status;
         }
+
+        if ($status) {
+            return 1;
+        } else {
+            return 0;
+        }
+    }
+
+    public function deleteImageItem($picture_id)
+    {
+        $picture = ReportsPict::find($picture_id);
+        if (!$picture) {
+            return 0;
+        }
+
+        File::delete($picture->picture);
+        $status = $picture->delete();
+
+        if ($status) {
+            return 1;
+        } else {
+            return 0;
+        }
+    }
+
+    public function updateImageItem(Request $request, $picture_id)
+    {
+        $this->validate($request, ['keterangan' => 'required']);
+
+        $picture = ReportsPict::find($picture_id);
+        if (!$picture) {
+            return 0;
+        }
+
+        $picture->keterangan = $request->keterangan;
+        $status = $picture->save();
+
+        if ($status) {
+            return 1;
+        } else {
+            return 0;
+        }
+    }
+
+    public function updateField(Request $request, $id)
+    {
+        $this->validate($request, [
+            'field' => 'required|in:desc,recomendation,jobdesc',
+            'value' => 'required',
+        ]);
+
+        $reports = Reports::find($id);
+        if (!$reports) {
+            return 0;
+        }
+
+        $reports->{$request->field} = $request->value;
+        $status = $reports->save();
+
         if ($status) {
             return 1;
         } else {
